Auto-crawl all known vhosts for host
This commit adds VHOST checks for each host in the autocrawler
execution cycle. All hostnames recorded as ntype VHOST for the IP
being scanned are added, as well as the hosts DB name if it
matches a regex, stripping special chars from the beginning in the
event that the name is recorded as *.domain.tld.
Jobwaiting modified to use Rex::ThreadSafe.sleep() as well.